pyproject.toml
[project]
name = "reachy-agent"
version = "0.1.0"
description = "Embodied AI agent for Reachy Mini robot"
requires-python = ">=3.10"
authors = [
    {name = "Claude in the Shell Team"}
]
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "anthropic>=0.30.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "structlog>=24.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "fastmcp>=0.1.0",
]

[project.optional-dependencies]
voice = [
    "pyaudio>=0.2.13",
    "openwakeword>=0.6.0",
    "websockets>=12.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1",
    "black>=24.0",
    "ruff>=0.3.0",
    "mypy>=1.8",
    "types-PyYAML>=6.0",
]

[project.scripts]
reachy-agent = "reachy_agent.__main__:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
